pyproject.toml
[tool.poetry]
name = "project-fast-curso"
version = "0.1.0"
description = ""
authors = ["Luis Henrique <luishenrique15.lh4@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"
pytest = "^8.2.2"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
